refactor(structure): Relocate scylla/orchestrator.py into scylla/e2e/
Move EvalOrchestrator and OrchestratorConfig from the orphaned
scylla/orchestrator.py into scylla/e2e/orchestrator.py, aligning it
with the consistent sub-package structure used by all other modules.

- Copy scylla/orchestrator.py → scylla/e2e/orchestrator.py (no content changes)
- Update scylla/cli/main.py to import from scylla.e2e.orchestrator
- Update tests/unit/e2e/test_orchestrator.py to import from scylla.e2e.orchestrator
- Add re-exports of EvalOrchestrator and OrchestratorConfig to scylla/e2e/__init__.py
- Remove stale 'orchestrator' entry from scylla/__init__.__all__
- Delete scylla/orchestrator.py via git rm
- Update CLAUDE.md repository structure section

All 2436 tests pass with 74.15% coverage.

Closes #847
